Decompose Dialog: move inferred schema fetch logic to primary cta click handler
