+++ src/components/editor/Bindings/SchemaInference/Dialog.tsx
@@ -11,12 +11,7 @@ import {
 import MessageWithLink from 'components/content/MessageWithLink';
 import InferenceDiffEditor from 'components/editor/Bindings/SchemaInference/Dialog/DiffEditor';
 import UpdateSchemaButton from 'components/editor/Bindings/SchemaInference/Dialog/UpdateSchemaButton';
-import {
-    useBindingsEditorStore_documentsRead,
-    useBindingsEditorStore_setDocumentsRead,
-    useBindingsEditorStore_setInferredSpec,
-    useBindingsEditorStore_setLoadingInferredSchema,
-} from 'components/editor/Bindings/Store/hooks';
+import { useBindingsEditorStore_documentsRead } from 'components/editor/Bindings/Store/hooks';
 import { CollectionData } from 'components/editor/Bindings/types';
 import AlertBox from 'components/shared/AlertBox';
 import {
@@ -24,16 +19,9 @@ import {
     secondaryButtonBackground,
     secondaryButtonHoverBackground,
 } from 'context/Theme';
-import { isEmpty } from 'lodash';
 import { Dispatch, SetStateAction, useState } from 'react';
 import { FormattedMessage } from 'react-intl';
-import { useEffectOnce, useLocalStorage, useUnmountPromise } from 'react-use';
-import getInferredSchema from 'services/schema-inference';
 import { useResourceConfig_currentCollection } from 'stores/ResourceConfig/hooks';
-import {
-    getStoredGatewayAuthConfig,
-    LocalStorageKeys,
-} from 'utils/localStorage-utils';
 
 interface Props {
     collectionData: CollectionData;
@@ -55,78 +43,12 @@ function SchemaInferenceDialog({
     const theme = useTheme();
 
     // Bindings Editor Store
-    const setInferredSpec = useBindingsEditorStore_setInferredSpec();
-
     const documentsRead = useBindingsEditorStore_documentsRead();
-    const setDocumentsRead = useBindingsEditorStore_setDocumentsRead();
-
-    const setLoadingInferredSchema =
-        useBindingsEditorStore_setLoadingInferredSchema();
 
     // Resource Config Store
     const currentCollection = useResourceConfig_currentCollection();
 
     const [schemaUpdateErrored] = useState<boolean>(false);
-
-    const [gatewayConfig] = useLocalStorage(
-        LocalStorageKeys.GATEWAY,
-        getStoredGatewayAuthConfig()
-    );
-
-    const resolveWhileMounted = useUnmountPromise();
-
-    useEffectOnce(() => {
-        if (currentCollection && gatewayConfig?.gateway_url) {
-            resolveWhileMounted(
-                getInferredSchema(gatewayConfig, currentCollection)
-            )
-                .then(
-                    (response) => {
-                        if (Object.hasOwn(collectionData.spec, 'schema')) {
-                            const { schema, ...additionalSpecKeys } =
-                                collectionData.spec;
-
-                            setInferredSpec(
-                                !isEmpty(response.schema)
-                                    ? {
-                                          writeSchema:
-                                              collectionData.spec.schema,
-                                          readSchema: response.schema,
-                                          ...additionalSpecKeys,
-                                      }
-                                    : null
-                            );
-                        } else if (
-                            Object.hasOwn(collectionData.spec, 'writeSchema')
-                        ) {
-                            const { writeSchema, ...additionalSpecKeys } =
-                                collectionData.spec;
-
-                            setInferredSpec(
-                                !isEmpty(response.schema)
-                                    ? {
-                                          writeSchema:
-                                              collectionData.spec.writeSchema,
-                                          readSchema: response.schema,
-                                          ...additionalSpecKeys,
-                                      }
-                                    : null
-                            );
-                        }
-
-                        setDocumentsRead(response.documents_read);
-                    },
-                    (error) => {
-                        setInferredSpec(error?.code === 404 ? null : undefined);
-
-                        setDocumentsRead(undefined);
-                    }
-                )
-                .finally(() => setLoadingInferredSchema(false));
-        } else {
-            setLoadingInferredSchema(false);
-        }
-    });
 
     const handlers = {
         closeConfirmationDialog: (event: React.MouseEvent<HTMLElement>) => {
